fix shutter test not execute engine tests
