+++ src/Nethermind/Nethermind.Merge.Plugin.Test/BaseEngineModuleTests.cs
@@ -1,10 +1,11 @@
-// SPDX-FileCopyrightText: 2022 Demerzel Solutions Limited
+// SPDX-FileCopyrightText: 2025 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using FluentAssertions;
 using Nethermind.Blockchain;
 using Nethermind.Blockchain.BeaconBlockRoot;
 using Nethermind.Blockchain.Blocks;
@@ -21,17 +22,20 @@ using Nethermind.Consensus.Withdrawals;
 using Nethermind.Core;
 using Nethermind.Core.Crypto;
 using Nethermind.Core.Events;
+using Nethermind.Core.Extensions;
 using Nethermind.Core.Specs;
 using Nethermind.Core.Test.Blockchain;
+using Nethermind.Core.Test.Builders;
 using Nethermind.Core.Timers;
 using Nethermind.Crypto;
 using Nethermind.Db;
-using Nethermind.Evm.Tracing;
 using Nethermind.Facade.Eth;
 using Nethermind.HealthChecks;
 using Nethermind.Int256;
+using Nethermind.JsonRpc;
 using Nethermind.Logging;
 using Nethermind.Merge.Plugin.BlockProduction;
+using Nethermind.Merge.Plugin.Data;
 using Nethermind.Merge.Plugin.GC;
 using Nethermind.Merge.Plugin.Handlers;
 using Nethermind.Merge.Plugin.Synchronization;
@@ -46,7 +50,7 @@ using NUnit.Framework;
 
 namespace Nethermind.Merge.Plugin.Test;
 
-public partial class EngineModuleTests
+public class BaseEngineModuleTests
 {
     [SetUp]
     public Task Setup()
@@ -148,6 +152,75 @@ public partial class EngineModuleTests
             chain.SpecProvider,
             new GCKeeper(NoGCStrategy.Instance, chain.LogManager),
             chain.LogManager);
+    }
+
+    protected async Task<IReadOnlyList<ExecutionPayload>> ProduceBranchV1(IEngineRpcModule rpc,
+        MergeTestBlockchain chain,
+        int count, ExecutionPayload startingParentBlock, bool setHead, Hash256? random = null,
+        ulong slotLength = 12)
+    {
+        List<ExecutionPayload> blocks = new();
+        ExecutionPayload parentBlock = startingParentBlock;
+        parentBlock.TryGetBlock(out Block? block);
+        UInt256? startingTotalDifficulty = block!.IsGenesis
+            ? block.Difficulty : chain.BlockFinder.FindHeader(block!.Header!.ParentHash!)!.TotalDifficulty;
+        BlockHeader parentHeader = block!.Header;
+        parentHeader.TotalDifficulty = startingTotalDifficulty +
+                                       parentHeader.Difficulty;
+        for (int i = 0; i < count; i++)
+        {
+            ExecutionPayload? getPayloadResult = await BuildAndGetPayloadOnBranch(rpc, chain, parentHeader,
+                parentBlock.Timestamp + slotLength,
+                random ?? TestItem.KeccakA, Address.Zero);
+            PayloadStatusV1 payloadStatusResponse = (await rpc.engine_newPayloadV1(getPayloadResult)).Data;
+            payloadStatusResponse.Status.Should().Be(PayloadStatus.Valid);
+            if (setHead)
+            {
+                Hash256 newHead = getPayloadResult!.BlockHash;
+                ForkchoiceStateV1 forkchoiceStateV1 = new(newHead, newHead, newHead);
+                ResultWrapper<ForkchoiceUpdatedV1Result> setHeadResponse = await rpc.engine_forkchoiceUpdatedV1(forkchoiceStateV1);
+                setHeadResponse.Data.PayloadStatus.Status.Should().Be(PayloadStatus.Valid);
+                setHeadResponse.Data.PayloadId.Should().Be(null);
+            }
+
+            blocks.Add((getPayloadResult));
+            parentBlock = getPayloadResult;
+            parentBlock.TryGetBlock(out block!);
+            block.Header.TotalDifficulty = parentHeader.TotalDifficulty + block.Header.Difficulty;
+            parentHeader = block.Header;
+        }
+
+        return blocks;
+    }
+
+    protected async Task<ExecutionPayload> BuildAndGetPayloadOnBranch(
+        IEngineRpcModule rpc, MergeTestBlockchain chain, BlockHeader parentHeader,
+        ulong timestamp, Hash256 random, Address feeRecipient)
+    {
+        PayloadAttributes payloadAttributes =
+            new() { Timestamp = timestamp, PrevRandao = random, SuggestedFeeRecipient = feeRecipient };
+
+        // we're using payloadService directly, because we can't use fcU for branch
+        string payloadId = chain.PayloadPreparationService!.StartPreparingPayload(parentHeader, payloadAttributes)!;
+
+        ResultWrapper<ExecutionPayload?> getPayloadResult =
+            await rpc.engine_getPayloadV1(Bytes.FromHexString(payloadId));
+        return getPayloadResult.Data!;
+    }
+
+    protected ExecutionPayload CreateParentBlockRequestOnHead(IBlockTree blockTree)
+    {
+        Block? head = blockTree.Head ?? throw new NotSupportedException();
+        return new ExecutionPayload()
+        {
+            BlockNumber = head.Number,
+            BlockHash = head.Hash!,
+            StateRoot = head.StateRoot!,
+            ReceiptsRoot = head.ReceiptsRoot!,
+            GasLimit = head.GasLimit,
+            Timestamp = head.Timestamp,
+            BaseFeePerGas = head.BaseFeePerGas,
+        };
     }
 
     public class MergeTestBlockchain : TestBlockchain
@@ -304,57 +377,5 @@ public partial class EngineModuleTests
 
         public async Task<MergeTestBlockchain> Build(ISpecProvider? specProvider = null) =>
             (MergeTestBlockchain)await Build(specProvider, null);
-    }
-}
-
-public class TestBlockProcessorInterceptor : IBlockProcessor
-{
-    private readonly IBlockProcessor _blockProcessorImplementation;
-    public int DelayMs { get; set; }
-    public Exception? ExceptionToThrow { get; set; }
-
-    public TestBlockProcessorInterceptor(IBlockProcessor baseBlockProcessor, int delayMs)
-    {
-        _blockProcessorImplementation = baseBlockProcessor;
-        DelayMs = delayMs;
-    }
-
-    public Block[] Process(Hash256 newBranchStateRoot, IReadOnlyList<Block> suggestedBlocks, ProcessingOptions processingOptions, IBlockTracer blockTracer)
-    {
-        if (DelayMs > 0)
-        {
-            Thread.Sleep(DelayMs);
-        }
-
-        if (ExceptionToThrow is not null)
-        {
-            throw ExceptionToThrow;
-        }
-
-        return _blockProcessorImplementation.Process(newBranchStateRoot, suggestedBlocks, processingOptions, blockTracer);
-    }
-
-    public event EventHandler<BlocksProcessingEventArgs>? BlocksProcessing
-    {
-        add => _blockProcessorImplementation.BlocksProcessing += value;
-        remove => _blockProcessorImplementation.BlocksProcessing -= value;
-    }
-
-    public event EventHandler<BlockEventArgs>? BlockProcessing
-    {
-        add => _blockProcessorImplementation.BlockProcessing += value;
-        remove => _blockProcessorImplementation.BlockProcessing -= value;
-    }
-
-    public event EventHandler<BlockProcessedEventArgs>? BlockProcessed
-    {
-        add => _blockProcessorImplementation.BlockProcessed += value;
-        remove => _blockProcessorImplementation.BlockProcessed -= value;
-    }
-
-    public event EventHandler<TxProcessedEventArgs>? TransactionProcessed
-    {
-        add => _blockProcessorImplementation.TransactionProcessed += value;
-        remove => _blockProcessorImplementation.TransactionProcessed -= value;
     }
 }
