+++ src/Nethermind/Nethermind.Merge.Plugin.Test/EngineModuleTests.V1.cs
@@ -1086,45 +1086,6 @@ public partial class EngineModuleTests
         }
     }
 
-    protected async Task<IReadOnlyList<ExecutionPayload>> ProduceBranchV1(IEngineRpcModule rpc,
-        MergeTestBlockchain chain,
-        int count, ExecutionPayload startingParentBlock, bool setHead, Hash256? random = null,
-        ulong slotLength = 12)
-    {
-        List<ExecutionPayload> blocks = new();
-        ExecutionPayload parentBlock = startingParentBlock;
-        parentBlock.TryGetBlock(out Block? block);
-        UInt256? startingTotalDifficulty = block!.IsGenesis
-            ? block.Difficulty : chain.BlockFinder.FindHeader(block!.Header!.ParentHash!)!.TotalDifficulty;
-        BlockHeader parentHeader = block!.Header;
-        parentHeader.TotalDifficulty = startingTotalDifficulty +
-                                       parentHeader.Difficulty;
-        for (int i = 0; i < count; i++)
-        {
-            ExecutionPayload? getPayloadResult = await BuildAndGetPayloadOnBranch(rpc, chain, parentHeader,
-                parentBlock.Timestamp + slotLength,
-                random ?? TestItem.KeccakA, Address.Zero);
-            PayloadStatusV1 payloadStatusResponse = (await rpc.engine_newPayloadV1(getPayloadResult)).Data;
-            payloadStatusResponse.Status.Should().Be(PayloadStatus.Valid);
-            if (setHead)
-            {
-                Hash256 newHead = getPayloadResult!.BlockHash;
-                ForkchoiceStateV1 forkchoiceStateV1 = new(newHead, newHead, newHead);
-                ResultWrapper<ForkchoiceUpdatedV1Result> setHeadResponse = await rpc.engine_forkchoiceUpdatedV1(forkchoiceStateV1);
-                setHeadResponse.Data.PayloadStatus.Status.Should().Be(PayloadStatus.Valid);
-                setHeadResponse.Data.PayloadId.Should().Be(null);
-            }
-
-            blocks.Add((getPayloadResult));
-            parentBlock = getPayloadResult;
-            parentBlock.TryGetBlock(out block!);
-            block.Header.TotalDifficulty = parentHeader.TotalDifficulty + block.Header.Difficulty;
-            parentHeader = block.Header;
-        }
-
-        return blocks;
-    }
-
     [Test]
     public async Task ExecutionPayloadV1_set_and_get_transactions_roundtrip()
     {
@@ -1231,21 +1192,6 @@ public partial class EngineModuleTests
             await rpc.engine_newPayloadV1(executionPayload);
         executePayloadResult.Data.Status.Should().Be(PayloadStatus.Valid);
         return executionPayload;
-    }
-
-    protected async Task<ExecutionPayload> BuildAndGetPayloadOnBranch(
-        IEngineRpcModule rpc, MergeTestBlockchain chain, BlockHeader parentHeader,
-        ulong timestamp, Hash256 random, Address feeRecipient)
-    {
-        PayloadAttributes payloadAttributes =
-            new() { Timestamp = timestamp, PrevRandao = random, SuggestedFeeRecipient = feeRecipient };
-
-        // we're using payloadService directly, because we can't use fcU for branch
-        string payloadId = chain.PayloadPreparationService!.StartPreparingPayload(parentHeader, payloadAttributes)!;
-
-        ResultWrapper<ExecutionPayload?> getPayloadResult =
-            await rpc.engine_getPayloadV1(Bytes.FromHexString(payloadId));
-        return getPayloadResult.Data!;
     }
