+++ src/Nethermind/Nethermind.Shutter.Test/ShutterBlockHandlerTests.cs
@@ -12,7 +12,7 @@ using Nethermind.Merge.Plugin.Test;
 namespace Nethermind.Shutter.Test;
 
 [TestFixture]
-class ShutterBlockHandlerTests : EngineModuleTests
+class ShutterBlockHandlerTests : BaseEngineModuleTests
 {
     [Test]
     public void Can_wait_for_valid_block()
